Kernel/USB: Move the USB components as a subfolder to the Bus directory
